FDI Analysis
About the project:
Investment is a game of understanding historic data of investment objects under different events but it is still a game of chances to minimize the risk we apply analytics to find the equilibrium investment. To understand the Foreign direct investment in India for the 17 years from 2000-01 to 2016-17. This dataset contains sector and financial year-wise data of FDI in India Sector-wise investment analysis Year-wise investment analysis.
TECHNOLOGY :
Business Intelligence
DOMAIN:
Finance